requirements.txt
aiogram>=3.0.0
python-dotenv
google-api-python-client
google-auth-httplib2
google-auth-oauthlib 
